+++ browser/modules/infoClick.js
@@ -1,6 +1,6 @@
 /*
  * @author     Martin Høgh <mh@mapcentia.com>
- * @copyright  2013-2018 MapCentia ApS
+ * @copyright  2013-2022 MapCentia ApS
  * @license    http://www.gnu.org/licenses/#AGPL  GNU AFFERO GENERAL PUBLIC LICENSE 3
  */
 
@@ -10,16 +10,16 @@ import {LAYER} from "./layerTree/constants";
 
 const MODULE_ID = `infoClick`;
 
-var cloud;
-var backboneEvents;
-var utils;
-var clicktimer;
-var sqlQuery;
-var layerTree;
-var _layers;
-var qstore = [];
-var active = false;
-var _self = false;
+let cloud;
+let backboneEvents;
+let utils;
+let clicktimer;
+let sqlQuery;
+let layerTree;
+let _layers;
+let qstore = [];
+let active = false;
+let _self = false;
 let blocked = false;
 let advancedInfo;
 
@@ -73,13 +73,13 @@ module.exports = {
             // Reset all SQL Query layers
             backboneEvents.get().trigger("sqlQuery:clear");
 
-            var event = new geocloud.clickEvent(e, cloud.get());
+            const event = new geocloud.clickEvent(e, cloud.get());
             if (clicktimer) {
                 clearTimeout(clicktimer);
             } else {
                 clicktimer = setTimeout(function () {
                     clicktimer = undefined;
-                    var coords = event.getCoordinate(), wkt;
+                    let coords = event.getCoordinate(), wkt;
                     wkt = "POINT(" + coords.x + " " + coords.y + ")";
 
                     // Cross Multi select disabled
@@ -122,17 +122,10 @@ module.exports = {
         });
     },
 
-    /**
-     *
-     */
     reset: function () {
         sqlQuery.reset(qstore);
     },
 
-    /**
-     *
-     * @param a {boolean}
-     */
     active: function (a) {
         if (!a) {
             this.reset();
@@ -162,7 +155,6 @@ module.exports = {
 
     on: () => {
         active = true;
-
     },
 
     off: () => {
@@ -170,5 +162,4 @@ module.exports = {
         utils.cursorStyle().reset();
     }
 };
-